better torrent handling
Former-commit-id: 47182a0e804042299ad49fe3c3de70e133790d3d [formerly fecab5c607cc82a4685d4921f3d78d8d2c389bf5] [formerly 76d4081e79bcb86819b15e2ae75d1a866f0cba25 [formerly 01281225a61e16d524761bde648d7128e8752517]]
Former-commit-id: ecaeb0e885e7b0d33caeca281f3fbb44c63c5bd9 [formerly 7284bc9e29b09ca9a750d55c429ecea440390589]
Former-commit-id: 08ee9d04c867ecfd0b3ac5a334414cb4c14c9083

+++ .zsh/06-alias.zsh
@@ -222,16 +222,20 @@ function resolve_file {
 }
 
 function ta {
-    tmp_list=/tmp/torr_list
-      sleep .2s
-      for i in $@; echo $i >> $tmp_list
-      while read line; do
-          file_name="$(resolve_file $line)"
-          base_name="$(basename $file_name)"
-          \mv $file_name ~/torrent/$base_name && \
-          transmission-remote-cli ~/torrent/$base_name > /dev/null &&
-          echo "$fg[magenta][$fg[blue]-->>$fg[magenta]] $fg[blue]$reset_color $fg[green]{$fg[purple]$base_name $fg[blue]added $fg[green]}"
-      done < $tmp_list
-      rm $tmp_list
-      file_name=
+    if [ -z $(pidof transmission-daemon)  ]; then
+        tmp_list=/tmp/torr_list
+        sleep .2s
+        for i in $@; echo $i >> $tmp_list
+        while read line; do
+            file_name="$(resolve_file $line)"
+            base_name="$(basename $file_name)"
+            \mv $file_name ~/torrent/$base_name && \
+            transmission-remote-cli ~/torrent/$base_name > /dev/null &&
+            echo "$fg[magenta][$fg[blue]-->>$fg[magenta]] $fg[blue]$reset_color $fg[green]{$fg[purple]$base_name $fg[blue]added $fg[green]}"
+        done < $tmp_list
+        rm $tmp_list
+        file_name=
+    else
+        transmission-daemon
+    fi
 }
